# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  version: 1.0.0
  description: |
    REST API for registering users, signing in, and managing authenticated users' todos.

    The current implementation stores users and todos in process memory. Data is reset
    whenever the server restarts. Todos are only returned to the user identified by the
    access token.

    The `GET /api/auth/view` endpoint is retained for compatibility with the current
    implementation. It is public and returns internal user records, including password
    hashes, so it must not be exposed in production.
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: "{protocol}://{host}:{port}"
    description: Configurable API server
    variables:
      protocol:
        default: http
        enum:
          - http
          - https
      host:
        default: localhost
      port:
        default: "5000"
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Log in a user
      description: Authenticate with email and password and receive a one-hour JWT access token.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: ada@example.com
              password: correct-horse-battery-staple
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                status: successful
                message: You have been successfully logged in
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
        '400':
          description: Email or password is missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                message: Email and password are required
        '401':
          description: The email or password is incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                message: Invalid email or password
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use the token returned by `POST /api/auth/login`.
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum:
            - error
          example: error
        message:
          type: string
          example: Title is required
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: ada@example.com
        password:
          type: string
          format: password
          example: correct-horse-battery-staple
    LoginResponse:
      type: object
      required:
        - status
        - message
        - token
      properties:
        status:
          type: string
          enum:
            - successful
        message:
          type: string
          example: You have been successfully logged in
        token:
          type: string
          description: JWT access token valid for one hour
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
...